OneDrive/Desktop/RabbittAi/sales-insight-automator/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
python-dotenv==1.1.*
python-multipart==0.0.*
pydantic[email]==2.11.*
pandas==2.2.*
openpyxl==3.1.*
google-genai>=1.0,<2.0
slowapi==0.1.*
matplotlib==3.10.*
gunicorn==23.*
